% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trajectory.R
\name{addSlingShotTrajectories}
\alias{addSlingShotTrajectories}
\title{Add a Slingshot Trajectories to an Seurat Project #NEW}
\usage{
addSlingShotTrajectories(
  object = NULL,
  cells = NULL,
  name = "SlingShot",
  useGroups = NULL,
  principalGroup = NULL,
  groupBy = NULL,
  embedding = NULL,
  seed = 1
)
}
\arguments{
\item{object}{An `Seurat Object` object.}

\item{name}{A string indicating the name of the fitted trajectory to be added in `meta.data`.}

\item{useGroups}{A character vector that is used to select a subset of groups by name from the designated `groupBy` column
in `meta.data`. This limits the groups used to identify trajectories.}

\item{principalGroup}{The principal group which represents the group that will be the starting point for all trajectories.}

\item{groupBy}{A string indicating the column name from `meta.data` that contains the cell group definitions used in
`useGroups` to constrain trajectory analysis.}

\item{embedding}{A string indicating the name of the `embedding` object from the `Seurat Object` that should be used for trajectory analysis.}

\item{seed}{A number to be used as the seed for random number generation for trajectory creation.}
}
\value{
seurat object add SlingShot Curve
}
\description{
This function will fit a supervised trajectory in a lower dimensional space that
can then be used for downstream analyses.
}
