% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agingscore.R
\name{ageScore}
\alias{ageScore}
\title{a function to caculate aging score}
\usage{
ageScore(
  seurat = NULL,
  geneSet = NULL,
  identColumn = NULL,
  plot = FALSE,
  jitter = FALSE
)
}
\arguments{
\item{seurat}{a seurat object.}

\item{geneSet}{a vector of gene names}

\item{identColumn}{which column in metadata as idents}

\item{plot}{whether show aging score plot}
}
\description{
a function to caculate aging score
}
