% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap.R
\name{ArchRHeatmap}
\alias{ArchRHeatmap}
\title{function for matrix to doheatmap in archr Style}
\usage{
ArchRHeatmap(
  mat = NULL,
  scale = FALSE,
  limits = c(min(mat), max(mat)),
  colData = NULL,
  color = paletteContinuous(set = "solarExtra", n = 100),
  clusterCols = TRUE,
  clusterRows = FALSE,
  labelCols = FALSE,
  labelRows = FALSE,
  colorMap = NULL,
  useRaster = TRUE,
  rasterQuality = 5,
  split = NULL,
  fontSizeRows = 10,
  fontSizeCols = 10,
  fontSizeLabels = 8,
  colAnnoPerRow = 4,
  showRowDendrogram = FALSE,
  showColDendrogram = FALSE,
  customRowLabel = NULL,
  customRowLabelIDs = NULL,
  customColLabel = NULL,
  customColLabelIDs = NULL,
  customLabelWidth = 0.75,
  rasterDevice = "png",
  padding = 45,
  borderColor = NA,
  draw = TRUE,
  name = "Heatmap"
)
}
\arguments{
\item{mat}{an matrix as sparse matrix}

\item{scale}{bool value,wether to do scale in heatmap.}

\item{limits}{limit the min and max value in mat.}

\item{colData}{provide cell bar in top heatmap if it is not null.}

\item{clusterCols}{same as clusterRows.}

\item{clusterRows}{bool value,whether do cluster in row.}

\item{customRowLabel}{custom label in row want to show.
...}

\item{seurat}{An `Seurat` object.}
}
\description{
function for matrix to doheatmap in archr Style
}
