% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pCR.R
\name{Combine10XContig}
\alias{Combine10XContig}
\title{combine 10X contig files}
\usage{
Combine10XContig(metaDF = NULL, pattern = "TCR")
}
\arguments{
\item{metaDF}{meta csv file,must include Path,Sample,Batch columns.Path is the path of filtered_contig_annotations.csv.}

\item{pattern}{contig is PCR or TCR}
}
\description{
combine 10X contig files
}
\examples{
 data("pCRexample")
Combine10XContig(pCRexample[["meta"]])
str(pCRexample[["combined"]])
}
