% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pCR.R
\name{contigWrapperSeurat}
\alias{contigWrapperSeurat}
\title{Adding clonotype information to a seurat or SCE object
This function adds the immune receptor information to the seurat or SCE object to the meta data. By defualt this function also
calculates the frequencies of the clonotypes by sequencing run
(groupBy = "none").  To change how the frequencies are calculated,
select a column header for the groupBy variable. Importantly,
before using combineExpression() ensure the barcodes of the seurat
or SCE object match the barcodes in the output of the
combinedContig() call. Check changeNames() to change the prefix of
the seurat object. If the dominant clonotypes have a greater
frequency than 500, adjust the cloneTypes variable.}
\usage{
contigWrapperSeurat(
  seurat = NULL,
  combined = NULL,
  groupby = "sample",
  cloneCall = "gene+nt"
)
}
\arguments{
\item{seurat}{seurat object}

\item{combined}{object come from Combine10XContig}

\item{groupby}{the column label in the combined contig object in which clonotype frequency will be calculated.}

\item{cloneCall}{How to call the clonotype - CDR3 gene (gene), CDR3 nucleotide (nt) CDR3 amino acid (aa), or CDR3 gene+nucleotide (gene+nt).}
}
\description{
Adding clonotype information to a seurat or SCE object
This function adds the immune receptor information to the seurat or SCE object to the meta data. By defualt this function also
calculates the frequencies of the clonotypes by sequencing run
(groupBy = "none").  To change how the frequencies are calculated,
select a column header for the groupBy variable. Importantly,
before using combineExpression() ensure the barcodes of the seurat
or SCE object match the barcodes in the output of the
combinedContig() call. Check changeNames() to change the prefix of
the seurat object. If the dominant clonotypes have a greater
frequency than 500, adjust the cloneTypes variable.
}
