% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellphonedb.R
\name{CPDBDotData}
\alias{CPDBDotData}
\title{function to prepare data from cellphone output for dot_plot}
\usage{
CPDBDotData(
  selected_rows = NULL,
  selected_columns = NULL,
  means_path = "./means.txt",
  pvalues_path = "./pvalues.txt",
  means_separator = "\\t",
  pvalues_separator = "\\t"
)
}
\arguments{
\item{selected_rows}{vector of row names in pvalue.txt' interacting_pair}

\item{selected_columns}{vector of col names in pvalue.txt}

\item{means_path}{the cellphone method analysis out's means.txt file}

\item{pvalues_path}{the cellphone method analysis out's pvalues.txt file}
}
\description{
function to prepare data from cellphone output for dot_plot
}
