% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellphonedb.R
\name{CPDBDotplot}
\alias{CPDBDotplot}
\title{dot plot for cellphone output}
\usage{
CPDBDotplot(
  selected_rows = NULL,
  selected_columns = NULL,
  filename = "plot.pdf",
  width = 8,
  height = 10,
  means_path = "./means.txt",
  pvalues_path = "./pvalues.txt",
  means_separator = "\\t",
  pvalues_separator = "\\t",
  output_extension = ".pdf"
)
}
\arguments{
\item{selected_rows}{vector of row names in pvalue.txt' interacting_pair}

\item{selected_columns}{vector of col names in pvalue.txt}

\item{filename}{plot filename}

\item{means_path}{the cellphone method analysis out's means.txt file}

\item{pvalues_path}{the cellphone method analysis out's pvalues.txt file}
}
\description{
dot plot for cellphone output
}
