% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellphonedb.R
\name{CPDBHeatmapData}
\alias{CPDBHeatmapData}
\title{function to prepare data from cellphone output for heatmap_plot}
\usage{
CPDBHeatmapData(
  meta_file = NULL,
  pvalues_file = NULL,
  meta_sep = "\\t",
  pvalues_sep = "\\t",
  pvalue = 0.05
)
}
\arguments{
\item{meta_file}{meta data file,cell_meta.txt}

\item{pvalues_file}{the cellphone method analysis out's pvalues.txt file}

\item{pvaue}{pvalue Threshold.}
}
\description{
function to prepare data from cellphone output for heatmap_plot
}
