% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellphonedb.R
\name{CPDBHeatmaps}
\alias{CPDBHeatmaps}
\title{heatmap plot for cellphone output}
\usage{
CPDBHeatmaps(
  meta_file = NULL,
  pvalues_file = NULL,
  count_filename = "count_heatmap.pdf",
  log_filename = "logcount_heatmap.pdf",
  count_network_filename = "count_network.txt",
  interaction_count_filename = "interaction_count.txt",
  count_network_separator = "\\t",
  interaction_count_separator = "\\t",
  show_rownames = T,
  show_colnames = T,
  scale = "none",
  cluster_cols = T,
  border_color = "white",
  cluster_rows = T,
  fontsize_row = 11,
  fontsize_col = 11,
  main = "",
  treeheight_row = 0,
  family = "Arial",
  treeheight_col = 0,
  col1 = "dodgerblue4",
  col2 = "peachpuff",
  col3 = "deeppink4",
  meta_sep = "\\t",
  pvalues_sep = "\\t",
  pvalue = 0.05
)
}
\arguments{
\item{meta_file}{meta data file,cell_meta.txt}

\item{pvalues_file}{the cellphone method analysis out's pvalues.txt file}

\item{pvaue}{pvalue Threshold.}
}
\description{
heatmap plot for cellphone output
}
