% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{create10xObj}
\alias{create10xObj}
\title{create seurat object from count matrix}
\usage{
create10xObj(
  file = NULL,
  species = "human",
  addPercent = TRUE,
  addID = NULL,
  preclust = TRUE
)
}
\arguments{
\item{file}{10X count or aggr filtered_feature_bc_matrix or filtered_feature_bc_matrix.h5}

\item{species}{human or others,if human,genes names upper,otherwise,title}

\item{addPercent}{whether add percent metrics}

\item{addID}{whther add ID name in barcode}

\item{preclust}{whether pre load analysis cluster}
}
\description{
create seurat object from count matrix
}
