% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CustomInteractionScore}
\alias{CustomInteractionScore}
\title{function to caculate Interaction score}
\usage{
CustomInteractionScore(
  seurat = NULL,
  column = NULL,
  idents = NULL,
  interaction_csv = NULL,
  slot = "data",
  eps = 1e-04
)
}
\arguments{
\item{seurat}{a seurat object.}

\item{column}{which column in metadata as idents.}

\item{idents}{which idents to be  used for scoring.}

\item{interaction_csv}{csv file,include two columns, left Ligand,right Receptor}

\item{slot}{slot name .}
}
\description{
function to caculate Interaction score
}
