% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof.R
\name{CytofToSeurat}
\alias{CytofToSeurat}
\title{function to convert cytof data into Seurat}
\usage{
CytofToSeurat(
  sample_csv = NULL,
  config_csv = NULL,
  N = 20000,
  useFeatures = NULL,
  batch_correct = FALSE,
  batch_key = "sample",
  write2count = FALSE,
  path2barcode10X = "3M-february-2018.txt",
  outdir = "cytof"
)
}
\arguments{
\item{sample_csv}{at least three columns(Path,Sample,Condition)}

\item{config_csv}{at least two columns(channel(Cd111Di,Nd143Di...),marker(CD8A,CD11C..))}

\item{N}{sample size,default sample 20000 each fcs file}

\item{useFeatures}{feature names provide to run scaleData and runpca.}

\item{batch_correct}{whether run batch correct to remove batch effect}

\item{batch_key}{when remove batch correct,which key use  to run.}

\item{path2barcode10X}{10X barcode file,can be found in cellranger path}
}
\description{
function to convert cytof data into Seurat
}
