% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEplot.R
\name{DAPlot}
\alias{DAPlot}
\title{Volcano Plot of Difference Analysis Markers}
\usage{
DAPlot(
  deDF = NULL,
  origin = "seurat",
  showGenes = NULL,
  cutoff = 0.5,
  x_limit = 5,
  y_limit = 0,
  down = "HC",
  up = "VKH",
  colorManual = c("skyblue", "gray", "red"),
  outdir = "Plots"
)
}
\arguments{
\item{deDF}{markers DataFrame}

\item{origin}{the markers from which platform,only support seurat and scanpy}

\item{showGenes}{the genew want to show in Volcano}

\item{cutoff}{the value to cutoff log2fc}

\item{x_limit}{the limits value for x_axis}

\item{y_limit}{the limits value for y_axis}

\item{down}{which value as down and show in left}

\item{up}{which value as up and show in right}

\item{colorManual}{color platte for scale_color_manual}

\item{outdir}{path to save plot}
}
\description{
Volcano Plot of Difference Analysis Markers
}
