% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEplot.R
\name{DAvenn}
\alias{DAvenn}
\title{venn or upset plot for different markers}
\usage{
DAvenn(
  deDF = NULL,
  origin = "seurat",
  plotAS = "venn",
  width = 8,
  height = 6,
  outdir = "Plots"
)
}
\arguments{
\item{deDF}{markers DataFrame}

\item{origin}{the markers from which platform,only support seurat and scanpy}

\item{plotAS}{venn or upset plot}

\item{width}{figure width}

\item{height}{figure height}

\item{outdir}{path to save plot}
}
\description{
venn or upset plot for different markers
}
