% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap.R
\name{DoLikeArchRHeatmap}
\alias{DoLikeArchRHeatmap}
\title{function doheatmap like DoHeatmap in seurat package}
\usage{
DoLikeArchRHeatmap(
  seurat = NULL,
  groupBy = NULL,
  N = 30,
  slot = "scale.data",
  scale = TRUE,
  limits = NULL,
  threads = 8
)
}
\arguments{
\item{seurat}{an seurat object.}

\item{groupBy}{which group used as Idents in seurat when find markers.}

\item{N}{the top number features to show in heatmap}

\item{slot}{which slot to be use in heatmap,default(scale.data).}

\item{scale}{whether do scale in heatmap.}

\item{limits}{limits the max and min value in heatmap.}

\item{threads}{the number of cores to be used.}
}
\description{
function doheatmap like DoHeatmap in seurat package
}
