% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSI.R
\name{.computeLSI}
\alias{.computeLSI}
\title{compute LSI(pass)}
\usage{
.computeLSI(
  mat = NULL,
  LSIMethod = 1,
  scaleTo = 10^4,
  nDimensions = 50,
  binarize = TRUE,
  outlierQuantiles = c(0.02, 0.98),
  seed = 1,
  verbose = FALSE
)
}
\description{
compute LSI(pass)
}
