% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEmbedding.R
\name{EmbPlot}
\alias{EmbPlot}
\title{Embeeding plot}
\usage{
EmbPlot(
  seurat = NULL,
  assay = "RNA",
  colorBy = NULL,
  features = NULL,
  embedding = "umap",
  outDir = NULL,
  imputeWeights = NULL,
  splitBy = NULL,
  combine = FALSE,
  width = 12,
  height = 10
)
}
\arguments{
\item{seurat}{an seurat object.}

\item{colorBy}{`meta.data` or `matrix`.}

\item{features}{feature names in seurat object}

\item{embedding}{ebedding name in seurat}
}
\description{
Embeeding plot
}
