% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellphonedb.R
\name{exportCellPhoneDB}
\alias{exportCellPhoneDB}
\title{a function to export count and metadata for cellphonedb}
\usage{
exportCellPhoneDB(
  seurat = NULL,
  cells = NULL,
  features = NULL,
  slot = "data",
  selectCol = "seurat_clusters",
  species = "human",
  outdir = "cellphoneDB",
  path2CPDB = findCellPhoneDB(),
  runCPDB = FALSE
)
}
\arguments{
\item{seurat}{a seurat object}

\item{cells}{feature names vector.}

\item{slot}{which slot to use.}

\item{selectCol}{which column to export as cell_type.}

\item{outdir}{path to save tables}

\item{runCPDB}{whether to run cellphonedb}

\item{cellphonedbPath}{cellphonedb path}
}
\description{
a function to export count and metadata for cellphonedb
}
