% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubletScore.R
\name{filterDoublets}
\alias{filterDoublets}
\title{function to filter doublet cells}
\usage{
filterDoublets(
  object = NULL,
  sampleCol = NULL,
  cutEnrich = 1,
  cutScore = -Inf,
  filterRatio = 1
)
}
\arguments{
\item{object}{a Seurat Object.}

\item{sampleCol}{which column in metadata as Sample}

\item{cutEnrich}{DoubletEnrichment value to cutoff}

\item{cutScore}{DoubletScore value to cutoff}
}
\description{
function to filter doublet cells
}
