% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{fixPlotSize}
\alias{fixPlotSize}
\title{set plot size}
\usage{
fixPlotSize(
  p = NULL,
  plotWidth = unit(6, "in"),
  plotHeight = unit(6, "in"),
  margin = 0.25,
  height = 1,
  it = 0.05,
  newPage = FALSE
)
}
\description{
set plot size
}
