% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getGroupPval}
\alias{getGroupPval}
\title{function to caculate genes in two status's differece pvalue,only support wilcox.test now}
\usage{
getGroupPval(seurat = NULL, geneSet = NULL, groupBy = NULL, slot = "data")
}
\arguments{
\item{seurat}{a seurat object.}

\item{geneSet}{a vector of gene names.}

\item{groupBy}{which column as condition(length ==2)}

\item{slot}{slot name,only support counts and data}
}
\description{
function to caculate genes in two status's differece pvalue,only support wilcox.test now
}
