% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trajectory.R
\name{getSeuratTrajectory}
\alias{getSeuratTrajectory}
\title{function to get Trajectory Object.}
\usage{
getSeuratTrajectory(
  seurat = NULL,
  splitBy = NULL,
  useFM = FALSE,
  name = "Trajectory",
  groupEvery = 1,
  log2Norm = TRUE,
  slot = "data",
  features = NULL,
  scaleTo = 10000,
  smoothWindow = 11,
  threads = 16
)
}
\arguments{
\item{seurat}{an `seurat` object.}

\item{splitBy}{column in `meta.data` use as ident for find marker or spliting for features.}

\item{useFM}{bool value,whether to use `FindMarkers`.}

\item{name}{the trajectory'name in `meta.data` after caculate specify trajectory.}

\item{slot}{slot name in seurat.}

\item{features}{list of features,each list include many genes.}
}
\description{
function to get Trajectory Object.
}
