% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGroups.R
\name{GgPlot}
\alias{GgPlot}
\title{function to plot feature Groups}
\usage{
GgPlot(
  seurat = NULL,
  assay = "RNA",
  features = NULL,
  outDir = NULL,
  splitBy = NULL,
  groupBy = NULL,
  combine = FALSE,
  width = 12,
  height = 10
)
}
\arguments{
\item{seurat}{an seurat object.}

\item{assay}{which assay to use.}

\item{features}{features in matrix}

\item{groupBy}{which column  in `meta.data` use as group.}

\item{combine}{whether combine plots}
}
\description{
function to plot feature Groups
}
