% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{IdentRename}
\alias{IdentRename}
\title{Rename one ident to a new name}
\usage{
IdentRename(obj = NULL, oldname = NULL, newname = NULL)
}
\arguments{
\item{obj}{seurat object}

\item{oldname}{old ident name you want to rename}

\item{newname}{new name you want to use}
}
\description{
Rename one ident to a new name
}
\examples{
x<-IdentRename(obj=pbmc_small,oldname = "0",newname = "C0")
table(Idents(x))
}
