% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImputeWeight.R
\name{ImputeWeights}
\alias{ImputeWeights}
\title{function to caculate impute weights}
\usage{
ImputeWeights(
  object = NULL,
  reducedDims = "IterativeLSI",
  dimsToUse = NULL,
  corCutOff = 0.75,
  td = 3,
  ka = 4,
  sampleCells = 5000,
  nRep = 2,
  k = 15,
  epsilon = 1,
  randomSuffix = FALSE,
  threads = 8,
  seed = 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{an seurat object.}

\item{reducedDims}{Embedding name in seurat object.}

\item{dimsToUse, integer}{number of dims to use.}

\item{sampleCells}{number of sample cells used}
}
\description{
function to caculate impute weights
}
