% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{makeMultObject}
\alias{makeMultObject}
\title{merge single seurat objects into multiple}
\usage{
makeMultObject(objs = NULL, samples = NULL, status = NULL, format = "seurat")
}
\arguments{
\item{objs}{seurat objects or seurat rds files}

\item{samples}{sample names relative objs}

\item{status}{statu names relative obj}

\item{format}{seurat or file}
}
\description{
merge single seurat objects into multiple
}
