% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetascapeUtils.R
\name{metascapeBarPlot}
\alias{metascapeBarPlot}
\title{metascape GO result left and right plot}
\usage{
metascapeBarPlot(
  leftDF = NULL,
  rightDF = NULL,
  pal = "Set1",
  save = TRUE,
  outdir = "metascape"
)
}
\arguments{
\item{leftDF}{DataFrame show in left}

\item{rightDF}{DataFrame show in right}

\item{pal}{which color pal use}
}
\description{
metascape GO result left and right plot
}
