% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImputeWeight.R
\name{MyimputeMatrix}
\alias{MyimputeMatrix}
\title{impute matrix use impute weights}
\usage{
MyimputeMatrix(mat = NULL, imputeWeights = NULL, threads = 16, verbose = FALSE)
}
\arguments{
\item{mat}{an matrix or sparse matrix.}

\item{imputeWeights}{impute weight object from function `ImputeWeights`.}
}
\description{
impute matrix use impute weights
}
