% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEmbedding.R
\name{MyplotEmbedding}
\alias{MyplotEmbedding}
\title{function to plot Embedding}
\usage{
MyplotEmbedding(
  seurat = NULL,
  assay = "RNA",
  embedding = "UMAP",
  colorBy = "metadata",
  name = "Sample",
  imputeWeights = NULL,
  pal = NULL,
  size = 0.1,
  sampleCells = NULL,
  highlightCells = NULL,
  rastr = TRUE,
  quantCut = c(0.01, 0.99),
  discreteSet = NULL,
  continuousSet = NULL,
  randomize = TRUE,
  keepAxis = FALSE,
  baseSize = 10,
  plotAs = NULL,
  labelSize = 0,
  ...
)
}
\arguments{
\item{seurat}{an seurat object.}

\item{assay}{which assay to use.}

\item{embedding}{which embedding to use plot.}

\item{colorBy}{`meta.data` or `matrix`.}

\item{name}{which column use for color by ,column in `meta.data` or feature in matrix}

\item{imputeWeights}{provide imputeWeights
...}
}
\description{
function to plot Embedding
}
