% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGroups.R
\name{MyplotGroups}
\alias{MyplotGroups}
\title{function to plot Groups}
\usage{
MyplotGroups(
  seurat = NULL,
  groupBy = "Sample",
  assay = "RNA",
  name = "Krt14",
  maxCells = 1000,
  quantCut = c(0.002, 0.998),
  pal = NULL,
  discreteSet = "stallion",
  ylim = NULL,
  size = 0.5,
  baseSize = 6,
  ratioYX = NULL,
  ridgeScale = 2,
  plotAs = "violin",
  ...
)
}
\arguments{
\item{seurat}{an seurat object.}

\item{groupBy}{which column  in `meta.data` use as group.}

\item{assay}{which assay to use.}

\item{name}{which feature in matrix}

\item{plotAS}{ridges or viloin plot.}
}
\description{
function to plot Groups
}
