% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{MyplotPDF}
\alias{MyplotPDF}
\title{function to save plotlist}
\usage{
MyplotPDF(
  ...,
  name = "Plot",
  width = 6,
  height = 6,
  outpath = NULL,
  addDOC = TRUE,
  useDingbats = FALSE,
  plotList = NULL
)
}
\arguments{
\item{name}{filename to save plot}

\item{width}{image width}

\item{height}{image height}

\item{outpath}{out dir to save image}

\item{addDOC}{whether to add date message}

\item{plotlist}{ggplot object list or complex heatmap object list}
}
\description{
function to save plotlist
}
