% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{objectPipeline}
\alias{objectPipeline}
\title{main pipeline for seurat object
you need to install SeuratWrappers,hamony first}
\usage{
objectPipeline(
  obj = NULL,
  nHVG = 2000,
  batch_correct = FALSE,
  batch_key = NULL,
  batch_method = "harmony",
  CellCycle = TRUE
)
}
\arguments{
\item{obj}{seurat object}

\item{nHVG}{deafult 2000,number of HVG}

\item{batch_correct}{whether batch correct}

\item{batch_key}{if batch correct,which key as batch}

\item{batch_method}{if batch correct,which method use}

\item{CellCycle}{whether do CellCycle to remove CellCycle effect}
}
\description{
main pipeline for seurat object
you need to install SeuratWrappers,hamony first
}
