% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGroups.R, R/utils.R
\name{quantileCut}
\alias{quantileCut}
\title{function to quantile cut}
\usage{
quantileCut(x = NULL, lo = 0.025, hi = 0.975, maxIf0 = TRUE)

quantileCut(x = NULL, lo = 0.025, hi = 0.975, maxIf0 = TRUE)
}
\description{
function to quantile cut

quantile cut
}
