% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{ReIdent}
\alias{ReIdent}
\title{re use column as Idents}
\usage{
ReIdent(obj = NULL, useCol = "seurat_clusters")
}
\arguments{
\item{obj}{seurat object}

\item{useCol}{which column as Idents}
}
\description{
re use column as Idents
}
