% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{saveImage}
\alias{saveImage}
\title{save image}
\usage{
saveImage(plotList, filename, width = 12, height = 16)
}
\arguments{
\item{plotList}{plot object list}

\item{filename}{filename to save the plot list}

\item{width}{image width}

\item{height}{image height}
}
\description{
save image
}
