% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellranger.R
\docType{data}
\name{scMatrix2CellRangerH5}
\alias{scMatrix2CellRangerH5}
\title{code to transform matrix into 10x h5 matrix}
\format{
An object of class \code{character} of length 1.
}
\usage{
scMatrix2CellRangerH5
}
\description{
code to transform matrix into 10x h5 matrix
}
\keyword{datasets}
