% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{setAppServer}
\alias{setAppServer}
\title{a simple Shiny App for Visualization}
\usage{
setAppServer(
  obj = NULL,
  browserTheme = "cosmo",
  host = "10.100.44.33",
  port = 5013,
  launch.browser = FALSE
)
}
\arguments{
\item{obj}{seurat object}

\item{browserTheme}{shiny browser theme}

\item{host}{host ip address}

\item{port}{ip port}

\item{launch.browser}{whether launch browser}
}
\description{
a simple Shiny App for Visualization
}
