% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Seurat2AnnData}
\alias{Seurat2AnnData}
\title{Convert Seurat into Anndata object}
\usage{
Seurat2AnnData(
  seurat = NULL,
  cells = cells,
  features = NULL,
  name = "seurat",
  outDir = "anndata"
)
}
\arguments{
\item{seurat}{seurat object}

\item{cells}{cells use to be exported}

\item{features}{features use to be exported}

\item{name}{prefix name for anndata file}

\item{outdir}{output path to save anndata}
}
\description{
Convert Seurat into Anndata object
}
