% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsplot.R
\name{statsBoxViolin}
\alias{statsBoxViolin}
\title{using ggstatsplot to visual gene expression or continous name in metadata}
\usage{
statsBoxViolin(
  obj,
  colorBy = "matrix",
  groupBy = "seurat_clusters",
  name = "CD8A",
  plot.type = "boxviolin",
  type = "parametric",
  slot = "data",
  p.adjust.method = "fdr",
  ...
)
}
\arguments{
\item{obj}{Seurat Object}

\item{colorBy}{matrix or metadata}

\item{groupBy}{which group in metadata as groupBy factor}

\item{name}{which variable will be plotting}

\item{plot.type}{which type to plot ,boxviolin,box or violin}

\item{type}{A character specifying the type of statistical approach. Four possible options:parametric,nonparametric,robust,bayes}

\item{slot}{which slot in Seurat Object to use,data ot count}

\item{p.adjust.method}{fdr(default),Adjustment method for p-values for multiple comparisons. Possible methods are: holm,hochberg, hommel, bonferroni, BH, BY,fdr, none.}

\item{...}{detail see ggstatsplot docs}
}
\value{
a ggplot object
}
\description{
using ggstatsplot to visual gene expression or continous name in metadata
}
