% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROGUE.R
\name{wrapperROGUE}
\alias{wrapperROGUE}
\title{Caculate ROGUE values for Cell Cluster and conclude their their heterogeneity.}
\usage{
wrapperROGUE(
  object = NULL,
  slot = "counts",
  useFeatures = NULL,
  draw = TRUE,
  ext = "pdf",
  k = NULL,
  auto = TRUE,
  platform = "UMI",
  labelCol = "label_fine",
  sampleCol = "Sample",
  outdir = NULL
)
}
\arguments{
\item{object}{Seurat or SingleCellExperiment or SummarizedExperiment Object}

\item{slot}{which slot data use,default counts}

\item{useFeatures}{features use to caculate rogue.value}

\item{draw}{whether draw plot}

\item{ext}{if draw,which figure style to save,default pdf}

\item{k}{value for caculate rogue}

\item{auto}{when k is NULL,will auto caculate k value}

\item{platform}{which platform,"UMI" or "full-length"}

\item{labelCol}{which column in metadata as labels(celltype)}

\item{sampleCol}{which column in metadata as sample}

\item{outdir}{path to save figures}
}
\value{
return as list of ent.res and rogue.res
}
\description{
Caculate ROGUE values for Cell Cluster and conclude their their heterogeneity.
}
