% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellranger.R
\name{generateCloupe}
\alias{generateCloupe}
\title{run cellranger reanalysis(only for cellranger version 3.1.0)}
\usage{
generateCloupe(
  crgPython = "python",
  targetMatrix = "outs/filtered_feature_bc_matrix",
  CRH5 = "filtered_feature_bc_matrix.h5",
  id = "reanalysis",
  path2CRG = "/home/ye/Software/cellranger-3.1.0"
)
}
\arguments{
\item{crgPython}{path of  cellranger python,eg(cellranger-3.1.0/miniconda-cr-cs/4.3.21-miniconda-cr-cs-c10/bin/python)}

\item{targetMatrix}{target matrix want to transform eg(outs/filtered_feature_bc_matrix)}

\item{id}{id name}

\item{path2CRG}{path of cellranger,only support cellranger-3.1.0}
}
\description{
run cellranger reanalysis(only for cellranger version 3.1.0)
}
\examples{
path2CRG="/home/test/Software/cellranger-3.1.0"
}
